+++ src/core/persistence/migrations.rs
@@ -1,2 +1,3 @@
 pub mod version_1;
 pub mod version_2;
+pub mod version_3;