+++ src/services/buienradar.rs
@@ -112,17 +112,11 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
         .find(|measurement| measurement.station_id == station_id)
         .ok_or_else(|| format_err!("station {} is not found", station_id))?;
     tx.send(
-        Composer::new(format!("{}::{}::name", service_id, station_id))
-            .type_(MessageType::ReadLogged)
-            .value(Value::Text(measurement.name.clone()))
-            .timestamp(measurement.timestamp)
-            .into(),
-    )?;
-    tx.send(
         Composer::new(format!("{}::{}::weather_description", service_id, station_id))
             .type_(MessageType::ReadLogged)
             .value(Value::Text(measurement.weather_description.clone()))
             .timestamp(measurement.timestamp)
+            .title(format!("{} Description", measurement.name))
             .into(),
     )?;
     if let Some(degrees) = measurement.temperature {
@@ -131,6 +125,7 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
                 .type_(MessageType::ReadLogged)
                 .value(ThermodynamicTemperature::new::<thermodynamic_temperature::degree_celsius>(degrees))
                 .timestamp(measurement.timestamp)
+                .title(format!("{} Temperature", measurement.name))
                 .into(),
         )?;
     }
@@ -140,6 +135,7 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
                 .type_(MessageType::ReadLogged)
                 .value(ThermodynamicTemperature::new::<thermodynamic_temperature::degree_celsius>(degrees))
                 .timestamp(measurement.timestamp)
+                .title(format!("{} Ground Temperature", measurement.name))
                 .into(),
         )?;
     }
@@ -149,6 +145,7 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
                 .type_(MessageType::ReadLogged)
                 .value(ThermodynamicTemperature::new::<thermodynamic_temperature::degree_celsius>(degrees))
                 .timestamp(measurement.timestamp)
+                .title(format!("{} Feel Temperature", measurement.name))
                 .into(),
         )?;
     }
@@ -158,6 +155,7 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
                 .type_(MessageType::ReadLogged)
                 .value(Value::Bft(bft))
                 .timestamp(measurement.timestamp)
+                .title(format!("{} Wind Force", measurement.name))
                 .into(),
         )?;
     }
@@ -167,6 +165,7 @@ fn send_readings(actual: BuienradarFeedActual, service_id: &str, station_id: u32
                 .type_(MessageType::ReadLogged)
                 .value(Value::WindDirection(point))
                 .timestamp(measurement.timestamp)
+                .title(format!("{} Wind Direction", measurement.name))
                 .into(),
         )?;
     }